requirements/base.txt
Django==2.2.4
django-autoslug==1.9.6
django-ckeditor==5.7.1
django-jet==1.0.8
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
Pillow==7.1.2
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pytils==0.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
